Add files via upload
Addressed bug that triggers with no resample factors (--rsf flag) omitted on command line call, or if any of the --rsf factors are identically 1. The bug is of the form:

...site-packages/tvtk/util/ctf.py:123: RuntimeWarning: invalid value encountered in double_scalars

Currently, a kludge fix has been put in place that forces the re-sampling to run, even with unit resample factors, which fixes the problem. Future updates will address the underlying issue with apparent ctf.py incompatability.